refactor(jcode-chat): split session store responsibilities
